[nrf fromtree] net: wifi_mgmt: change type of `wifi_scan_params->chan`
All WiFi channel numbers fit within a 8 bit number, as the maximum
allocated channel is 233. This halves the memory requirement.

(cherry picked from commit dd090f06b7ed45bdc88335082a10bbe0db2ce439)

+++ include/zephyr/net/wifi_utils.h
@@ -100,7 +100,7 @@ int wifi_utils_parse_scan_ssids(char *scan_ssids_str,
  * @retval -errno value in case of failure.
  */
 int wifi_utils_parse_scan_chan(char *scan_chan_str,
-			       uint16_t chan[][WIFI_CHANNEL_MAX]);
+			       uint8_t chan[][WIFI_CHANNEL_MAX]);
 
 /**
  * @}
